Fix: Tab to next tabIndex only works when dropdown is closed
Without this change, when pressing tab with the dropdown open it focuses on a random field as the dropdown has no tabIndex associated with it.
This change adds tabIndex to the dropdown so pressing tab with it open advances to the next ui-select.
Note that this only works for ui-selects as the tabIndex is actually multiplied by 2.